refactoring parts to their logical place
