+++ src/components/access-manager-layout/access-manager-body.tsx
@@ -1,6 +1,5 @@
 import React, { useEffect, useState } from 'react';
 import { MemberSettings } from '../member-settings/member-settings';
-import { AccessManagerNewMemberButton } from './access-manager-new-member-button';
 import { getMembersList } from '../../clients/get-members-list';
 import { Member } from '../../clients/mock-data';
 
@@ -26,7 +25,6 @@ export function AccessManagerBody() {
       {members.map(({ person_id, firstname, lastname }) => (
         <MemberSettings key={person_id} id={person_id} firstname={firstname} lastname={lastname} />
       ))}
-      <AccessManagerNewMemberButton />
     </>
   );
 }
